Transfer the functions to set up MPEC into estimation_mpec_functions.
